requirements.txt
aiohttp==3.7.4.post0
aiokafka==0.7.2
aiomysql==0.0.22
aiosqlite==0.17.0
aredis==1.1.8
asgiref==3.5.0
async-timeout==3.0.1
attrs==21.4.0
certifi==2021.10.8
chardet==4.0.0
click==8.0.3
easydict==1.9
fastapi==0.63.0
h11==0.13.0
idna==3.3
iso8601==0.1.16
jetengine==0.1.3
kafka-python==2.0.2
motor==1.3.1
multidict==6.0.2
numpy==1.21.5
pycrypto==2.6.1
pydantic==1.8.2
pymongo==3.6.0
PyMySQL==0.9.3
pypika-tortoise==0.1.3
pytz==2021.3
PyYAML==5.4.1
scipy==1.7.3
starlette==0.13.6
tortoise-orm==0.18.0
typing_extensions==4.1.1
uvicorn==0.15.0
yarl==1.7.2
opencv-python-headless==4.5.5.64